feat(PrintView): Allow multiple event selection for print
This change modifies the `PrintView` component to allow users to select multiple events for printing or exporting to Excel. Previously, only a single event or "all" could be selected.

The state `selectedEventIdForPrint` has been updated to `selectedEventIdsForPrint` (an array of strings) to accommodate this new functionality. An effect has been added to automatically select all available events when the component loads or when event data changes. New handler functions `handleEventSelectionChange` and `handleSelectAllEventsFor` are introduced to manage the selection logic.